Got tests running again.
