+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -30,7 +30,7 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
 
             extensions.configure<LibraryExtension> {
                 configureKotlinAndroid(this)
-                defaultConfig.targetSdk = 34
+                defaultConfig.targetSdk = 33
             }
 
             configurations.configureEach {
