+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/chatlist/ChatListItem.kt
@@ -1,10 +1,10 @@
 package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.chatlist
 
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.Post
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.PostPojo
 import kotlinx.datetime.LocalDate
 
 sealed class ChatListItem {
-    data class PostChatListItem(val post: Post) : ChatListItem()
+    data class PostChatListItem(val post: PostPojo) : ChatListItem()
 
     data class DateDivider(val localDate: LocalDate) : ChatListItem()
 }
