+++ feature/metis/conversation/src/test/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/EmptyVisibleMetisContextManager.kt
@@ -1,0 +1,10 @@
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation
+
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.visiblemetiscontextreporter.VisibleMetisContext
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.visiblemetiscontextreporter.VisibleMetisContextManager
+
+object EmptyVisibleMetisContextManager : VisibleMetisContextManager {
+    override fun registerMetisContext(metisContext: VisibleMetisContext) = Unit
+
+    override fun unregisterMetisContext(metisContext: VisibleMetisContext) = Unit
+}
