+++ feature/metis/manage-conversations/build.gradle.kts
@@ -18,5 +18,7 @@ dependencies {
 
     implementation(libs.androidx.paging.runtime)
     implementation(libs.androidx.paging.compose)
+
+    testImplementation(project(":feature:metis-test"))
 }
