+++ feature/metis/shared/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/shared/content/dto/AnswerPost.kt
@@ -1,7 +1,7 @@
 package de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto
 
 import de.tum.informatics.www1.artemis.native_app.core.model.account.User
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.AnswerPostDb
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.AnswerPostPojo
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -23,7 +23,7 @@ data class AnswerPost(
 
     override val serverPostId: Long = id ?: 0L
 
-    constructor(answerPostDb: AnswerPostDb, post: StandalonePost) : this(
+    constructor(answerPostDb: AnswerPostPojo, post: StandalonePost) : this(
         id = answerPostDb.serverPostId,
         author = User(id = answerPostDb.authorId),
         content = answerPostDb.content,
