--- feature/metis/src/test/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/overview/CreateConversationE2eTest.kt
@@ -1,148 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.overview
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.test.ExperimentalTestApi
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.hasAnyAncestor
-import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.hasText
-import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.performTextInput
-import androidx.lifecycle.SavedStateHandle
-import de.tum.informatics.www1.artemis.native_app.core.common.test.EndToEndTest
-import de.tum.informatics.www1.artemis.native_app.core.data.filterSuccess
-import de.tum.informatics.www1.artemis.native_app.core.model.account.User
-import de.tum.informatics.www1.artemis.native_app.core.common.test.DefaultTestTimeoutMillis
-import de.tum.informatics.www1.artemis.native_app.core.test.test_setup.DefaultTimeoutMillis
-import de.tum.informatics.www1.artemis.native_app.core.common.test.testServerUrl
-import de.tum.informatics.www1.artemis.native_app.feature.login.test.user2DisplayName
-import de.tum.informatics.www1.artemis.native_app.feature.login.test.user2Username
-import de.tum.informatics.www1.artemis.native_app.feature.login.test.user3Username
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ConversationBaseTest
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.Conversation
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.GroupChat
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.OneToOneChat
-import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.service.network.getConversation
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ui.conversation.create_personal_conversation.CreatePersonalConversationScreen
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ui.conversation.create_personal_conversation.CreatePersonalConversationViewModel
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ui.conversation.create_personal_conversation.TEST_TAG_CREATE_PERSONAL_CONVERSATION_BUTTON
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ui.conversation.member_selection.MemberSelectionBaseViewModel
-import de.tum.informatics.www1.artemis.native_app.feature.metis.ui.conversation.member_selection.TEST_TAG_MEMBER_SELECTION_SEARCH_FIELD
-import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.first
-import org.junit.Test
-import org.junit.experimental.categories.Category
-import org.junit.runner.RunWith
-import org.koin.test.get
-import org.robolectric.RobolectricTestRunner
-import kotlin.test.assertIs
-
-@Category(EndToEndTest::class)
-@RunWith(RobolectricTestRunner::class)
-class CreateConversationE2eTest : ConversationBaseTest() {
-
-    /**
-     * Tests that upon entering the name of user2 their name will be displayed in the potential recipients list.
-     */
-    @OptIn(ExperimentalTestApi::class)
-    @Test(timeout = DefaultTestTimeoutMillis)
-    fun `displays users on search`() {
-        val viewModel = setupUiAndViewModel()
-
-        composeTestRule
-            .onNodeWithTag(TEST_TAG_MEMBER_SELECTION_SEARCH_FIELD)
-            .performTextInput(user2DisplayName)
-
-        testDispatcher.scheduler.advanceTimeBy(MemberSelectionBaseViewModel.QUERY_DEBOUNCE_TIME * 2)
-
-        runBlockingWithTestTimeout {
-            viewModel.potentialRecipients.filterSuccess().filter { it.isNotEmpty() }.first()
-        }
-
-        composeTestRule.waitUntilAtLeastOneExists(
-            hasText(user2DisplayName, substring = true) and !hasAnyAncestor(
-                hasTestTag(
-                    TEST_TAG_MEMBER_SELECTION_SEARCH_FIELD
-                )
-            ),
-            DefaultTimeoutMillis
-        )
-    }
-
-    @Test(timeout = DefaultTestTimeoutMillis)
-    fun `can create personal conversation`() {
-        val createdConversation = createConversationTestImpl(
-            listOf(User(username = user2Username))
-        )
-
-        assertIs<de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.OneToOneChat>(createdConversation, "Created conversation is not a OneToOneChat")
-    }
-
-    @Test(timeout = DefaultTestTimeoutMillis)
-    fun `can create group conversation`() {
-        val createdConversation = createConversationTestImpl(
-            listOf(
-                User(username = user2Username),
-                User(username = user3Username)
-            )
-        )
-
-        assertIs<de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.GroupChat>(createdConversation, "Created conversation is not a group conversation")
-    }
-
-    private fun createConversationTestImpl(recipients: List<User>): de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.Conversation {
-        var createdConversationId: Long? = null
-
-        val viewModel = setupUiAndViewModel(
-            onConversationCreated = { conversationId ->
-                createdConversationId = conversationId
-            }
-        )
-
-        recipients.forEach(viewModel::addRecipient)
-
-        composeTestRule
-            .onNodeWithTag(TEST_TAG_CREATE_PERSONAL_CONVERSATION_BUTTON)
-            .assertIsDisplayed()
-            .performClick()
-
-        composeTestRule.waitUntil(DefaultTimeoutMillis) { createdConversationId != null }
-
-        return runBlockingWithTestTimeout {
-            conversationService.getConversation(
-                course.id!!,
-                createdConversationId!!,
-                accessToken,
-                testServerUrl
-            )
-                .orThrow("Could not load created conversation")
-        }
-    }
-
-    private fun setupUiAndViewModel(
-        onConversationCreated: (Long) -> Unit = {}
-    ): CreatePersonalConversationViewModel {
-        val viewModel = CreatePersonalConversationViewModel(
-            courseId = course.id!!,
-            conversationService = get(),
-            accountService = get(),
-            serverConfigurationService = get(),
-            networkStatusProvider = get(),
-            savedStateHandle = SavedStateHandle(),
-            coroutineContext = testDispatcher
-        )
-
-        composeTestRule.setContent {
-            CreatePersonalConversationScreen(
-                modifier = Modifier.fillMaxSize(),
-                viewModel = viewModel,
-                onConversationCreated = onConversationCreated,
-                onNavigateBack = {}
-            )
-        }
-
-        return viewModel
-    }
-}
